Btn Show Borrows (deocamdata afiseaza doar id-urile)

+++ src/main/java/Repositories/BorrowRepository.java
@@ -1,22 +1,27 @@
 package Repositories;
 
-
 import Entities.Borrow;
 import jakarta.annotation.security.DeclareRoles;
+import jakarta.ejb.Stateless;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
+
 import java.util.List;
 
+@Stateless
 @DeclareRoles({"administrator", "manager", "user"})
 public class BorrowRepository {
 
     @PersistenceContext
     private EntityManager entityManager;
 
+
+
     public List<Borrow> getAllBorrows() {
         return entityManager.createQuery("SELECT b FROM Borrow b", Borrow.class)
                 .getResultList();
     }
+
 
     public Borrow getBorrowById(int borrowId) {
         return entityManager.find(Borrow.class, borrowId);
@@ -32,7 +37,14 @@ public class BorrowRepository {
 
 
 
-    public void deleteBorrow(Borrow borrow) {
-        entityManager.remove(borrow);
+    public void deleteBorrow(Borrow borrowId) {
+        Borrow borrow = entityManager.find(Borrow.class, borrowId);
+        if (borrow != null) {
+            entityManager.remove(borrow);
+        }
+    }
+
+    public void createBorrow(Borrow borrow) {
+        entityManager.persist(borrow);
     }
 }
